+++ src/components/location-finder/HeroImage/HeroImage.test.js
@@ -1,18 +1,18 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { axe, toHaveNoViolations } from 'jest-axe';
-import BottomPromo from './BottomPromo';
+import HeroImage from './HeroImage';
 
 expect.extend(toHaveNoViolations);
 
-describe('BottomPromo', () => {
-  test('renders BottomPromo component', () => {
-    render(<BottomPromo />);
+describe('HeroImage', () => {
+  test('renders HeroImage component', () => {
+    render(<HeroImage />);
   });
 });
 
 it('should not have basic accessibility issues', async () => {
-  const { container } = render(<BottomPromo />);
+  const { container } = render(<HeroImage />);
   const results = await axe(container);
   expect(results).toHaveNoViolations();
 });